Add a method to start a build with collected changes
This can be used by the startBuild implementations. If the scheduler
has collected and classified any changes, it will use
`addBuildsetForChanges` and flush the change classifications. If there
are no changes it will use `addBuildsetForSourceStampsWithDefaults`.
If `onlyIfChanged` was given and there are no important changes, the
method will return without triggering any builds. Collected changes
won't be flushed from the database.